+++ R/scheduler.R
@@ -35,16 +35,12 @@
 #' \dontrun{
 #' # create a Dockerfile of your script
 #' if(!require(containeRit)){
-#'   devtools::install_github("o2r-project/containerit")
+#'   remotes::install_github("o2r-project/containerit")
 #'   library(containeRit)
 #' }
 #' 
-#' script <- system.file("schedulescripts", "schedule.R", package = "googleComputeEngineR")
 #' 
-#' ## put the "schedule.R" script in the working directory
-#' file.copy(script, getwd())
-#' 
-#' 
+#' ## create your scheduled script, example below named schedule.R
 #' ## it will run the script whilst making the dockerfile
 #' container <- dockerfile("schedule.R",
 #'                         copy = "script_dir",