+++ R/template.R
@@ -1,66 +1,3 @@
-#' create the cloud_init file to upload
-#' @keywords internal
-#' @import assertthat
-get_cloud_init_file <- function(template,
-                                username = NULL, 
-                                password = NULL, 
-                                dynamic_image = NULL) {
-  
-  
-  if(template == "rstudio"){
-    assert_that(
-      is.string(username),
-      is.string(password)
-    )
-    
-  } else if(template == "dynamic"){
-    assert_that(
-      is.string(dynamic_image)
-    )
-  }
-  
-  switch(template,
-         rstudio = build_cloud_init_file_rstudio(template_file = template, 
-                                                 docker_image = "rocker/tidyverse", 
-                                                 dynamic_image = dynamic_image,
-                                                 username = username, 
-                                                 password = password),
-         dynamic = build_cloud_init_file(template_file = template, 
-                                         docker_image = dynamic_image, 
-                                         dynamic_image = dynamic_image),
-         shiny = build_cloud_init_file(template_file = template, 
-                                       docker_image = "rocker/shiny", 
-                                       dynamic_image = dynamic_image),
-         opencpu = build_cloud_init_file(template_file = template, 
-                                         docker_image = "opencpu/base", 
-                                         dynamic_image = dynamic_image),
-         "r-base" = build_cloud_init_file(template_file = template, 
-                                          docker_image = "rocker/r-base", 
-                                          dynamic_image = dynamic_image),
-         ropensci = build_cloud_init_file_rstudio(template_file = template, 
-                                                  docker_image = "rocker/ropensci:dev", 
-                                                  dynamic_image = dynamic_image,
-                                                  username = username,
-                                                  password = password))
-}
-
-# build a cloud init file for all non-rstudio 
-build_cloud_init_file <- function(template_file, docker_image, dynamic_image){
-  cloud_init <- get_template_file(template_file)
-  cloud_init_file <- readChar(cloud_init, nchars = file.info(cloud_init)$size)
-  image <- get_image(docker_image, dynamic_image = dynamic_image)
-  sprintf(cloud_init_file, image)
-}
-
-# build the cloud-init file with a username and password
-build_cloud_init_file_rstudio <- function(template_file, docker_image, dynamic_image, username, password){
-  cloud_init <- get_template_file(template_file)
-  cloud_init_file <- readChar(cloud_init, nchars = file.info(cloud_init)$size)
-  image <- get_image(docker_image, dynamic_image = dynamic_image)
-  sprintf(cloud_init_file, username, password, username, image)
-}
-
-
 #' Create a template container VM
 #' 
 #' This lets you specify templates for the VM you want to launch
@@ -73,7 +10,8 @@ build_cloud_init_file_rstudio <- function(template_file, docker_image, dynamic_i
 #' @param username username if needed (RStudio)
 #' @param password password if needed (RStudio)
 #' @param image_family An image-family.  It must come from the \code{cos-cloud} family.
-#' @param dynamic_image Supply an alternative to the default Docker image here to download
+#' @param dynamic_image Supply an alternative to the default Docker image for the template
+#' @param wait Whether to wait for the VM to launch before returning. Default \code{TRUE}.
 #' @inheritDotParams gce_vm_container
 #' 
 #' @details 
@@ -82,20 +20,20 @@ build_cloud_init_file_rstudio <- function(template_file, docker_image, dynamic_i
 #' 
 #' \itemize{
 #'   \item rstudio An RStudio server docker image with tidyverse and devtools
+#'   \item rstudio-gpu An RStudio server with popular R machine learning libraries and GPU driver.  Will launch a GPU enabled VM.
+#'   \item rstudio-shiny An RStudio server with Shiny also installed, proxied to /shiny
 #'   \item shiny A Shiny docker image
 #'   \item opencpu An OpenCPU docker image
 #'   \item r_base Latest version of R stable
-#'   \item ropensci RStudio and tidyverse with all ropensci packages on CRAN and Github
-#'   \item example A non-R test container running busybox
-#'   \item dynamic Supply your own docker image to download such as \code{rocker/verse}
+#'   \item dynamic Supply your own docker image within dynamic_image
 #'  }
 #'  
-#'  For \code{dynamic} templates you will need to launch the docker image with any ports you want opened, 
+#' For \code{dynamic} templates you will need to launch the docker image with any ports you want opened, 
 #'    other settings etc. via \link{docker_run}.
 #' 
 #' Use \code{dynamic_image} to override the default rocker images e.g. \code{rocker/shiny} for shiny, etc. 
 #'  
-#' @return The VM object
+#' @return The VM object, or the VM startup operation if \code{wait=FALSE}
 #' 
 #' @examples 
 #' 
@@ -114,49 +52,88 @@ build_cloud_init_file_rstudio <- function(template_file, docker_image, dynamic_i
 #' }
 #' 
 #' 
-#' @export  
-gce_vm_template <- function(template = c("rstudio","shiny","opencpu",
-                                         "r-base", "example",
-                                         "dynamic", "ropensci"),
+#' @export
+#' @import assertthat
+gce_vm_template <- function(template = c("rstudio",
+                                         "shiny",
+                                         "opencpu",
+                                         "r-base",
+                                         "dynamic",
+                                         "rstudio-gpu", 
+                                         "rstudio-shiny"),
                             username=NULL,
                             password=NULL,
                             dynamic_image=NULL,
                             image_family = "cos-stable",
+                            wait = TRUE,
                             ...){
   
+  template <- match.arg(template)
+  assert_that(is.flag(wait),
+              is.string(image_family))
   dots <- list(...)
   
   if(is.null(dots$name)){
     message("No VM name specified, defaulting to ", template)
     dots$name <- template
   }
-  
-  template <- match.arg(template)
-
-  cloud_init_file <- get_cloud_init_file(template,
-                                         username = username, 
-                                         password = password, 
-                                         dynamic_image = dynamic_image)
-  
-  ## metadata
-  upload_meta <- list(template = template)
-  if(grepl("rstudio", template)){
-    upload_meta$rstudio_users <- username
+  if(is.null(dots$zone)){
+    dots$zone <- gce_get_global_zone()
+  }
+  if(is.null(dots$project)){
+    dots$project <- gce_get_global_project()
   }
   
+  # hack for gpu until nvidia-docker is supported on cos-cloud images
+  if(grepl("gpu$", template)){
+    # setup GPU specific options
+    dots            <- set_gpu_template(dots)
+    ss_file         <- get_template_file(template, "startupscripts")
+    startup_script  <- readChar(ss_file, nchars = file.info(ss_file)$size)
+    cloud_init_file <- NULL
+    image_family    <- "tf-latest-cu92"
+    image_project   <- "deeplearning-platform-release"
+  } else {
+    # creates cloud-config file that will call the startup script
+    cloud_init_file <- read_cloud_init_file(template)
+    startup_script  <- NULL
+    image_project   <-  "cos-cloud"
+  }
+
+  # adds metadata startup script will read
+  dots <- setup_shell_metadata(dots,
+                               template = template,
+                               username = username, 
+                               password = password, 
+                               dynamic_image = dynamic_image)
+
+  ## metadata
+  dots <- modify_metadata(dots, list(template = template,
+                                     "google-logging-enabled" = "true"))
+  
+  ## tag for http, shiny etc.
+  dots$tags <- template_tags(template)
+
   ## build VM
   job <- do.call(gce_vm_container,
-                 c(dots, list(
-                   cloud_init = cloud_init_file,
-                   image_family = image_family,
-                   tags = list(items = list("http-server")), # no use now
-                   metadata = upload_meta
-                 )))
+                 args = c(list(
+                     cloud_init = cloud_init_file,
+                     shell_script = startup_script,
+                     image_family = image_family,
+                     image_project = image_project
+                   ), 
+                   dots)
+                 )
   
-  gce_wait(job, wait = 10)
-  
-  ins <- gce_get_instance(dots$name)
-  ip <- gce_get_external_ip(dots$name)
+  if(wait){
+    gce_wait(job, wait = 10)
+  } else {
+    myMessage("Returning the VM startup job, not the VM instance.", level = 2)
+    return(job)
+  }
+
+  ins <- gce_get_instance(dots$name, project = dots$project, zone = dots$zone)
+  ip  <- gce_get_external_ip(ins, verbose = FALSE)
   
   ## where to find application
   ip_suffix <- ""
@@ -164,27 +141,29 @@ gce_vm_template <- function(template = c("rstudio","shiny","opencpu",
     ip_suffix <- "/ocpu/"
   }
 
-  myMessage("## VM ", paste0(template, " running at ", ip,ip_suffix), level = 3)
-  myMessage("Wait a few minutes for inital docker container to 
-            download and install before logging in.", level = 3)
+  myMessage("## VM Template: '", paste0(template, "' running at http://", ip,ip_suffix), level = 3)
+  myMessage("On first boot, wait a few minutes for docker container to install before logging in.", level = 3)
   
+  print(ins)
   ins
   
 }
 
-get_image <- function(default_image, dynamic_image = NULL){
-  ## override default rocker image
-  if(is.null(dynamic_image)){
-    image <- default_image
-  } else {
-    image <- dynamic_image
-  }
-  image
+template_tags <- function(template){
+  switch(template,
+         "rstudio" = list(items = list("http-server", "rstudio")),
+         "rstudio-gpu" = list(items = list("http-server","rstudio")),
+         "rstudio-shiny" = list(items = list("http-server","rstudio","shiny")),
+         "shiny" = list(items = list("http-server","shiny")),
+         "opencpu" = list(items = list("http-server","opencpu")),
+         "r-base" = list(items = list("r-base"))
+  )
 }
 
 #' Show the cloud-config template files
 #' 
-#' @param template
+#' @param template The template for the file to fetch
+#' @param folder Which folder within the package to fetch it for
 #' 
 #' This returns the file location of template files for use in \link{gce_vm_template}
 #' 
@@ -202,10 +181,12 @@ get_image <- function(default_image, dynamic_image = NULL){
 #'  }
 #' 
 #' @return file location
-#' @export
-get_template_file <- function(template){
+#' @noRd
+get_template_file <- function(template, type = c("cloudconfig", "startupscripts")){
+  type <- match.arg(type)
+  ext <- if(type == "cloudconfig") ".yaml" else ".sh"
   
-  system.file("cloudconfig", paste0(template,".yaml"), package = "googleComputeEngineR")
+  system.file(type, paste0(template, ext), package = "googleComputeEngineR")
   
 }